[PATCH] serial: MRi MRI-PCIDS1 dual port serial card
Add support for the MRi PCIDS1 dual port serial card.  This card is a
little controversial since it is the subject of a PCI vendor/device ID
clash.  (See
http://www.ussg.iu.edu/hypermail/linux/kernel/0303.1/0516.html).  I have
for now just used the hex ID 0x950a.  The divisor was part calculated part
iterated, so may not be exactly correct (but works for me at all settings
between 300 - 115300 bps).

Cc: Russell King <rmk@arm.linux.org.uk>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>