[PATCH] ipmi: OEM flag handling and hacks for some Dell machines
The ipmi driver does not have a way to handle firmware-generated events
which have the OEM[012] Data Available flags set.  In such a case, the
SMS_ATN bit may never get cleared by firmware, leaving the driver looping
infinitely but never able to make any progress.

This patch first simplifies storage and use of the data returned from an
IPMI Get Device ID command.

It then creates a new per-OEM handler hook, which should know how to handle
events with the OEM[012] Data Available flags set.  It then uses this to
implement a workaround for IPMI 1.5-capable Dell PowerEdge servers which
are susceptable to setting the OEM[012] Data Available flags when the
driver can't handle it.

Signed-off-by: Matt Domsch <Matt_Domsch@dell.com>
Signed-off-by: Corey Minyard <minyard@acm.org>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>