[PATCH] HPET, Maintainers
This patch identifies the HPET Maintainers.  Clemens in taking over as
primary maintainer for the HPET driver.  Clemens has i386 hardware with
HPET and is a better choice than me because of this.  I've shared this
patch with all cc: recipients and there is agreement on ownership.
Hopefully this eliminates future confusion in terms of where HPET
maintenance is owned.

Signed-off-by: Bob Picco <bob.picco@hp.com>
Cc: <clemens@ladisch.de>
Cc: <venkatesh.pallipadi@intel.com>
Cc: <bob.picco@hp.com>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>