[PATCH] RE: [PATCH 1/1] net driver: Add support for SMSC LAN911x line of ethernet chips
>
> The patch was badly wordwrapped.  Please fix and resend.
>

OK, I've fixed the wrapping and removed the CONFIG_ARM restriction.  I've also did my
best to modify the C style to conform to the comments.

I noticed that the patch is getting ignored by majordomo due to its size >100K.
Should it be broken up somehow to allow posting to the lists?

Signed-off-by: Dustin McIntire <dustin@sensoria.com>
Signed-off-by: Jeff Garzik <jeff@garzik.org>