kbuild/mkuboot.sh: allow spaces in CROSS_COMPILE
I'm currently using CROSS_COMPILE="ccache arm-linux-".  With that the bash
builtin command "type" searches for ccache and arm-linux-mkimage and so sets
MKIMAGE="/path/to/ccache" as I don't have arm-linux-mkimage.  Then the script
dies with an error, that ccache doesn't support the argument -A.

This patch adds some quoting such that it works again for me.

Please note that this patch doesn't help you if you use ${CROSSCOMPILE}-mkimage
and ccache as mkuboot.sh now searches for the command
"ccache arm-linux-mkimage".

Signed-off-by: Uwe Kleine-König <ukleinek@informatik.uni-freiburg.de>
Signed-off-by: Sam Ravnborg <sam@ravnborg.org>