ACPI: thinkpad-acpi: fix sysfs paths in documentation
The documentation used "thinkpad-acpi" to refer to the directories in
sysfs, while it should have been using "thinkpad_acpi".  Thanks to Hugh
Dickins for the error report.

I wish I could just call the module and everything else by the proper
name with the "-", instead of using these ugly translations to "_".

Signed-off-by: Henrique de Moraes Holschuh <hmh@hmh.eng.br>
Cc: Hugh Dickins <hugh@veritas.com>
Signed-off-by: Len Brown <len.brown@intel.com>