V4L/DVB (5985): Fix the min/max frequencies of some DVB-C frontends
The min frequencies of the DVB-C frontends are wrong. 

In Europe, the center frequency of the lowest channel is 50.5MHz and not 
51MHz. All known cards with the stv0297/tda0002x/ves1820 frontend are 
able to tune to this frequency.

I've changed the range to the lowest channel - 1/2 bandwidth and the 
highest channel + 1/2 bandwidth. For the design of the dvb driver, the 
frequency ranges must be part of the tuner and not of the frontend 
itself. The same frontend may be used for different tuners. 

The attached patch does only fix the ranges and not the design.

Signed-off-by: Hartmut Birr <e9hack@googlemail.com>
Signed-off-by: Oliver Endriss <o.endriss@gmx.de>
Signed-off-by: Mauro Carvalho Chehab <mchehab@infradead.org>