[SCSI] zfcp: update MAINTAINERS
we are planning a major rewrite of the zfcp driver,
meaning that a lot of patches will hit the mailing-list in the near future.

Since I can't support this additional work-load along with my other
responsibilities we are shifting the maintainership to
Christof Schmitt as the maintainer and
Martin Peschke as the co-maintainer.

Please support the two in providing us a new and more stable
zfcp environment.

Thanks
Swen

Signed-off-by: James Bottomley <James.Bottomley@HansenPartnership.com>