powerpc: udbg-based backend for hvc_console
This adds a new backend for the hvc console based on the low-level
udbg callbacks.  This effectively implements a working runtime console
in terms of the simple udbg primitives.  This is kind of a hack -
since udbg isn't something you really want to be using routinely - but
it's really useful during bringup.

This can be used to quickly implement a userspace-usable console while
you're working on a proper driver for whatever console I/O device the
hardware has.  Or, it can be used to avoid writing a full blown
tty/console driver entirely for quick-and-dirty I/O hardware that will
later be replaced by something else.

Signed-off-by: David Gibson <david@gibson.dropbear.id.au>
Signed-off-by: Paul Mackerras <paulus@samba.org>